+++ hswidget/animationwidget/widgetdcc.go
@@ -2,6 +2,8 @@ package animationwidget
 
 import (
 	"fmt"
+	"image"
+	"image/color"
 	"log"
 	"time"
 
@@ -9,12 +11,8 @@ import (
 	"github.com/ianling/imgui-go"
 
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2fileformats/d2dcc"
-	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2interface"
-	"github.com/OpenDiablo2/dialog"
 
-	"github.com/OpenDiablo2/HellSpawner/hscommon"
 	"github.com/OpenDiablo2/HellSpawner/hscommon/hsutil"
-	"github.com/OpenDiablo2/HellSpawner/hswidget"
 )
 
 const (
@@ -22,50 +20,12 @@ const (
 )
 
 type DccWidget struct {
-	widget *Widget
-	dcc    *d2dcc.DCC
+	*Widget
+	dcc *d2dcc.DCC
 }
 
-func (p *DccWidget) exportGif(state *DccWidgetState) error {
-	fpd := int32(p.dcc.FramesPerDirection)
-	firstFrame := state.controls.direction * fpd
-	images := state.images[firstFrame : firstFrame+fpd]
-
-	err := hsutil.ExportToGif(images, state.tickTime)
-	if err != nil {
-		return fmt.Errorf("error creating gif file: %w", err)
-	}
-
-	return nil
-}
-
-func (w *DccWidget) makePlayerLayout(state *DccWidgetState) giu.Layout {
-	playModeList := make([]string, 0)
-	for i := playModeForward; i <= playModePingPong; i++ {
-		playModeList = append(playModeList, i.String())
-	}
-
-	pm := int32(state.playMode)
-
-	return giu.Layout{
-		giu.Row(
-			giu.Checkbox("Loop##"+w.widget.id+"PlayRepeat", &state.repeat),
-			giu.Combo("##"+w.widget.id+"PlayModeList", playModeList[state.playMode], playModeList, &pm).OnChange(func() {
-				state.playMode = animationPlayMode(pm)
-			}).Size(comboW),
-			giu.InputInt("Tick time##"+w.widget.id+"PlayTickTime", &state.tickTime).Size(inputIntW).OnChange(func() {
-				state.ticker.Reset(time.Second * time.Duration(state.tickTime) / miliseconds)
-			}),
-			hswidget.PlayPauseButton("##"+w.widget.id+"PlayPauseAnimation", &state.isPlaying, w.widget.textureLoader).
-				Size(playPauseButtonSize, playPauseButtonSize),
-			giu.Button("Export GIF##"+w.widget.id+"exportGif").OnClick(func() {
-				err := w.exportGif(state)
-				if err != nil {
-					dialog.Message(err.Error()).Error()
-				}
-			}),
-		),
-	}
+func (w *DccWidget) getDcImage() DcImage {
+	return w.dcc
 }
 
 // Build build a widget
@@ -124,29 +84,171 @@ func (w *DccWidget) Build() {
 			imgui.EndGroup()
 		}),
 		giu.Separator(),
-		w.makePlayerLayout(viewerState),
+		makePlayerLayout(w, viewerState),
 		giu.Separator(),
 		widget,
 	}.Build()
 }
 
-func CreateDccWidget(tl hscommon.TextureLoader, state []byte, palette *[256]d2interface.Color, id string, dcc *d2dcc.DCC) giu.Widget {
-	widget := &Widget{
-		id: id,
-		palette: palette,
-		textureLoader: tl,
-	}
-
+func createDccWidget(state []byte, widget *Widget, dcc *d2dcc.DCC) giu.Widget {
 	dccWidget := &DccWidget{
-		widget: widget,
-		dcc: dcc,
+		Widget: widget,
+		dcc:    dcc,
 	}
 
-	if giu.Context.GetState(dccWidget.widget.getStateID()) == nil && state != nil {
+	if giu.Context.GetState(dccWidget.Widget.getStateID()) == nil && state != nil {
 		s := dccWidget.getState()
 		s.Decode(state)
 		dccWidget.setState(s)
 	}
 
 	return dccWidget
+}
+
+func (w *DccWidget) getState() *DccWidgetState {
+	var state *DccWidgetState
+
+	s := giu.Context.GetState(w.Widget.getStateID())
+
+	if s != nil {
+		state = s.(*DccWidgetState)
+	} else {
+		w.initState()
+		state = w.getState()
+	}
+
+	return state
+}
+
+func (w *DccWidget) setState(s giu.Disposable) {
+	giu.Context.SetState(w.Widget.getStateID(), s)
+}
+
+func (w *DccWidget) runPlayer(state *DccWidgetState) {
+	for range state.ticker.C {
+		if !state.isPlaying {
+			continue
+		}
+
+		numFrames := int32(w.dcc.FramesPerDirection - 1)
+		isLastFrame := state.controls.frame == numFrames
+
+		// update play direction
+		switch state.playMode {
+		case playModeForward:
+			state.isForward = true
+		case playModeBackward:
+			state.isForward = false
+		case playModePingPong:
+			if isLastFrame || state.controls.frame == 0 {
+				state.isForward = !state.isForward
+			}
+		}
+
+		// now update the frame number
+		if state.isForward {
+			state.controls.frame++
+		} else {
+			state.controls.frame--
+		}
+
+		state.controls.frame = int32(hsutil.Wrap(int(state.controls.frame), w.dcc.FramesPerDirection))
+
+		// next, check for stopping/repeat
+		isStoppingFrame := (state.controls.frame == 0) || (state.controls.frame == numFrames)
+
+		if isStoppingFrame && !state.repeat {
+			state.isPlaying = false
+		}
+	}
+}
+
+func (w *DccWidget) makeImagePixel(val byte) color.RGBA {
+	alpha := maxAlpha
+
+	if val == 0 {
+		alpha = 0
+	}
+
+	var r, g, b uint8
+
+	if w.Widget.palette != nil {
+		col := w.Widget.palette[val]
+		r, g, b = col.R(), col.G(), col.B()
+	} else {
+		r, g, b = val, val, val
+	}
+
+	RGBAColor := color.RGBA{
+		R: r,
+		G: g,
+		B: b,
+		A: alpha,
+	}
+
+	return RGBAColor
+}
+
+func (w *DccWidget) initState() {
+	// Prevent multiple invocation to LoadImage.
+	state := &DccWidgetState{
+		WidgetState: WidgetState{
+			isPlaying: false,
+			repeat:    false,
+			tickTime:  defaultTickTime,
+			playMode:  playModeForward,
+		},
+	}
+
+	state.ticker = time.NewTicker(time.Second * time.Duration(state.tickTime) / miliseconds)
+
+	w.setState(state)
+
+	go w.runPlayer(state)
+
+	totalFrames := w.dcc.NumberOfDirections * w.dcc.FramesPerDirection
+	state.images = make([]*image.RGBA, totalFrames)
+
+	for dirIdx := range w.dcc.Directions {
+		fw := w.dcc.Directions[dirIdx].Box.Width
+		fh := w.dcc.Directions[dirIdx].Box.Height
+
+		for frameIdx := range w.dcc.Directions[dirIdx].Frames {
+			absoluteFrameIdx := (dirIdx * w.dcc.FramesPerDirection) + frameIdx
+
+			frame := w.dcc.Directions[dirIdx].Frames[frameIdx]
+			pixels := frame.PixelData
+
+			state.images[absoluteFrameIdx] = image.NewRGBA(image.Rect(0, 0, fw, fh))
+
+			for y := 0; y < fh; y++ {
+				for x := 0; x < fw; x++ {
+					idx := x + (y * fw)
+					if idx >= len(pixels) {
+						continue
+					}
+
+					val := pixels[idx]
+
+					RGBAColor := w.makeImagePixel(val)
+					state.images[absoluteFrameIdx].Set(x, y, RGBAColor)
+				}
+			}
+		}
+	}
+
+	go func() {
+		textures := make([]*giu.Texture, totalFrames)
+
+		for frameIndex := 0; frameIndex < totalFrames; frameIndex++ {
+			frameIndex := frameIndex
+			w.Widget.textureLoader.CreateTextureFromARGB(state.images[frameIndex], func(t *giu.Texture) {
+				textures[frameIndex] = t
+			})
+		}
+
+		s := w.getState()
+		s.textures = textures
+		w.setState(s)
+	}()
 }
